requirements.txt
aniso8601==9.0.1
attrs==21.2.0
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==3.0.4
charset-normalizer==2.0.12
click==7.1.2
cryptography==3.4.7
defusedxml==0.7.1
dnspython==2.1.0
docutils==0.17.1
elasticsearch==7.13.4
Flask==1.1.4
flask-restx==0.5.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
googletrans==4.0.0rc1
greenlet==1.1.0
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
icalendar==4.0.7
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
lxml==4.8.0
MarkupSafe==2.0.1
numpy==1.21.3
pandas==1.3.4
paramiko==2.7.2
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
PyNaCl==1.4.0
pyrsistent==0.18.0
python-dateutil==2.8.2
python-docx==0.8.11
python-gvm==21.6.0
pytz==2021.1
requests==2.27.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.21
urllib3==1.26.6
Werkzeug==1.0.1
WTForms==3.0.1
